Based on patch from lxs, with some changes:
Add several new gcc warning flags, used in the KfM build process.

Put declarations before code.

Fix a bunch of signed/unsigned type mixes, mostly by changing variable
types to unsigned int.

Fix constness in handling name of default ccache name.

Make sure functions get declared with prototypes:
krb5int_pthread_loaded krb5int_gmt_mktime krb5int_aes_encrypt
krb5int_aes_decrypt gssint_mecherrmap_init gssint_mecherramp_get.

Don't shadow global names: stat accept index open encrypt.  Fix
variable shadowing in LDAP ASN.1 support.

Don't define unused krb5int_local_addresses.
Don't export internal krb5_change_set_password.
Fix error return indications from gssint_oid_to_mech.

Create and use k5-gmt_mktime.h to provide one global declaration of
krb5int_gmt_mktime, needed before we've generated krb5.h on some
platforms.

Not incorporated from initial patch: const changes in function
signatures.

ticket: 6096
status: open

git-svn-id: svn://anonsvn.mit.edu/krb5/trunk@20697 dc483132-0cff-0310-8789-dd5450dbe970

+++ src/lib/krb5/asn.1/asn1buf.h
@@ -22,14 +22,14 @@ int asn1buf_size
    ? 0 \
    : ((buf)->bound - (buf)->base + 1))
 
-int asn1buf_free
+unsigned int asn1buf_free
 	(const asn1buf *buf);
 /* requires  *buf is allocated
    effects   Returns the number of unused, allocated octets in *buf. */
 #define asn1buf_free(buf) \
   (((buf) == NULL || (buf)->base == NULL) \
-   ? 0 \
-   : ((buf)->bound - (buf)->next + 1))
+   ? 0U \
+   : (unsigned int)((buf)->bound - (buf)->next + 1))
 
 
 asn1_error_code asn1buf_ensure_space
